Update app to solve linter errors
