+++ app.rb
@@ -8,16 +8,20 @@ require 'json'
 require 'date'
 require_relative './src/genre'
 require_relative './src/music_album'
+require_relative './handler'
 
-class App
+class App < Handler
   def initialize
+    super
+    @genres = []
+    @music_albums = []
+
+    # @storage = Storage.new(self)
     @store = PreserveData.new
     @books = File.empty?('./store/books.json') ? [] : @store.load_data('./store/books.json')
     @labels = File.empty?('./store/labels.json') ? [] : @store.load_data('./store/labels.json')
     @games = File.empty?('./store/games.json') ? [] : @store.load_data('./store/games.json')
     @authors = File.empty?('./store/authors.json') ? [] : @store.load_data('./store/authors.json')
-    @genre = File.empty?('./store/genre.json') ? [] : @store.load_data('./store/genre.json')
-    @music_albums = File.empty?('./store/music_albums.json') ? [] : @store.load_data('./store/music_albums.json')
   end
 
   def list_books
@@ -94,7 +98,7 @@ class App
       puts 'List of Authors:'
       puts '=============================================:'
       puts 'ID - First Name - Last Name'
-      @authors.each do |author|
+      authors.each do |author|
         puts "#{author['id']} - #{author['first_name']} - #{author['last_name']}"
       end
       puts '=============================================:'
@@ -115,8 +119,8 @@ class App
     puts 'Enter last name'
     last_name = gets.chomp
     author = create_author(first_name, last_name)
-    @authors << author
-    @store.save_data(@authors, './store/authors.json')
+    authors << author
+    @store.save_data(authors, './store/authors.json')
     game = create_game_object(game_name, last_played_at, publish_date, multiplayer)
     @games << game
     @store.save_data(@games, './store/games.json')
@@ -141,69 +145,6 @@ class App
       'publish_date' => game.publish_date,
       'multiplayer' => game.multiplayer
     }
-  end
-
-  def add_music_album
-    puts 'Album name: '
-    name = gets.chomp
-    puts 'Genre: '
-    genre_name = gets.chomp
-    @genres.push(Genre.new(genre_name))
-    store_genre_data
-
-    puts 'Date of publish [Enter date in format (yyyy-mm-dd)]'
-    publish_date = gets.chomp
-
-    puts 'Is it available on Spotify? Y/N'
-    on_spotify = gets.chomp.downcase
-    case on_spotify
-    when 'y'
-      @music_albums.push(MusicAlbum.new(name, publish_date, true))
-    when 'n'
-      @music_albums.push(MusicAlbum.new(name, publish_date, false))
-    end
-    puts 'Music album created'
-    store_music_data
-  end
-
-  def albums
-    music_data = get_data('./store/music_albums.json')
-    puts 'No music found' if music_data.empty?
-    music_data.each do |music_album|
-      puts "Album_name: #{music_album['music_name']} | On_spotify: #{music_album['music_on_spotify']}"
-    end
-  end
-
-  def genres
-    data = get_data('./store/genres.json')
-    puts 'No genre found' if data.empty?
-    data.each do |genre|
-      puts "Genre name: #{genre['genre_name']}"
-    end
-  end
-
-  def store_music_data
-    array = []
-    @music_albums.each do |music|
-      array << {
-        music_id: music.id,
-        music_name: music.name,
-        music_on_spotify: music.on_spotify
-
-      }
-    end
-    update_data(array, './store/music_albums.json')
-  end
-
-  def store_genre_data
-    array = []
-    @genres.each do |genre|
-      array << {
-        genre_id: genre.id,
-        genre_name: genre.name
-      }
-    end
-    update_data(array, './store/genres.json')
   end
 
   def exit_app
